# src/main/resources/swagger/ms-cuentaBancaria_swagger.yaml
openapi: 3.0.1
info:
  title: Microservicio de Cuentas Bancarias
  version: 1.0.0
  description: |
    Microservicio para gestionar y administrar cuentas bancarias de clientes, permitiendo crear, obtener, actualizar y eliminar cuentas, así como realizar movimientos como depósitos y retiros.
  termsOfService: https://www.example.com/terms
  contact:
    email: vertkle@gmail.com
  license:
    name: Licencia MIT
    url: https://www.example.com/license
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /movement/deposit:
    post:
      operationId: createMovementDeposit
      summary: Crear depósito
      tags:
        - Movimientos Cuenta Bancaria
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Movement'
      responses:
        '201':
          description: Depósito creado exitosamente
        '400':
          description: Error en Request
        '404':
          description: Recurso no encontrado
  /movement/transfer:
    post:
      operationId: createTransfer
      summary: Crear transferencia
      tags:
        - Movimientos Cuenta Bancaria
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Movement'
      responses:
        '201':
          description: Transferencia creada exitosamente
        '400':
          description: Error en Request
        '404':
          description: Recurso no encontrado
  /movement/withdraw:
    post:
      operationId: createWithDraw
      summary: Crear retiro
      tags:
        - Movimientos Cuenta Bancaria
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Movement'
      responses:
        '201':
          description: Retiro creado exitosamente
        '400':
          description: Error en Request
        '404':
          description: Recurso no encontrado
  /movement/{account_id}:
    get:
      operationId: getMovementsByAccountId
      summary: Obtener movimientos de cuenta bancaria
      tags:
        - Movimientos Cuenta Bancaria
      parameters:
        - name: account_id
          in: path
          description: ID de la cuenta bancaria
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de movimientos de cuenta bancaria obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movement'
        '400':
          description: Error en request
        '404':
          description: Recurso no encontrado
...
  /debit_card:
    get:
      operationId: getAllDebitCard
      summary: Obtener todas las tarjetas de débito
      tags:
        - Tarjeta de débito
      responses:
        '200':
          description: Lista de tarjetas de debito, obtenidas exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DebitCard"
    post:
      operationId: createDebitCard
      summary: Agregar tarjeta de credito
      tags:
        - Tarjeta de débito
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebitCard'
      responses:
        '201':
          description: Tarjeta de débito creada exitosamente
        '400':
          description: Error en Request
    put:
      operationId: updateDebitCard
      summary: Actualizar Tarjeta débito
      tags:
        - Tarjeta de débito
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebitCard'
      responses:
        '200':
          description: Tarjeta de débito actualizada exitosamente
        '400':
          description: Error en Request
        '404':
          description: Recurso no encontrado
  /debit_card/{debit_card_id}:
    get:
      operationId: getDebitCardById
      summary: Obtener detalle de una tarjeta de débito
      tags:
        - Tarjeta de débito
      parameters:
        - name: debit_card_id
          in: path
          description: ID de la tarjeta de débito
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalle de tarjeta de débito obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebitCard'
        '400':
          description: Error en request
        '404':
          description: Recurso no encontrado
    delete:
      operationId: deleteDebitCard
      summary: Eliminar Tarjeta de débito
      tags:
        - Tarjeta de débito
      parameters:
        - name: debit_card_id
          in: path
          description: ID de la tarjeta de débito
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tarjeta de débito eliminada exitosamente
        '400':
          description: Error en request
        '404':
          description: Recurso no encontrado
...
components:
  schemas:
...
    Movement:
      type: object
      properties:
        _id:
          type: string
        account_id:
          type: string
        type_movement:
          type: string
        mount:
          type: number
        fecha:
          type: string
        fee:
          type: number
        destination:
          type: string
...
    AccountsSecundary:
      type: object
      properties:
        priority:
          type: number
        accountId:
          type: string
    DebitCard:
      type: object
      properties:
        customerId:
          type: string
        createdCardDebit:
          type: string
        numberCard:
          type: string
        expiration:
          type: string
        cvv2:
          type: string
        password:
          type: string
        accountIdPrincipal:
          type: string
        accountsSecundary:
          type: array
          items:
            $ref: "#/components/schemas/AccountsSecundary"
...